Create a coverage chart showing sampling rate

Only applicable to non-reactive sensors with "continuous" sampling

coverage(
  db,
  participant_id,
  sensor = "All",
  frequency,
  relative = TRUE,
  offset = "None",
  startDate = NULL,
  endDate = NULL
)

Arguments

db
A valid database connection. Schema must be that as it is created by open_db.

participant_id
A character string of one participant ID.

sensor
A character vector containing one or multiple sensors. See sensors for a list of available sensors. Use "All" for all available sensors.

frequency
A named numeric vector with sensors as names and the number of expected samples per hour

relative
Show absolute number of measurements or relative to the expected number? Logical value.

offset
Currently not used.

startDate
A date (or convertible to a date using as.Date) indicating the earliest date to show. Leave empty for all data. Must be used with endDate.

endDate
A date (or convertible to a date using as.Date) indicating the latest date to show.Leave empty for all data. Must be used with startDate

Value

Examples

setwd("~/data")

#> Error in setwd("~/data"): cannot change working directory

fix_json()

#> No JSON files found.

unzip()

#> Error in unzip(): argument "zipfile" is missing, with no default

freq <- c(
   Accelerometer = 720, # Once per 5 seconds. Can have multiple measurements.
   AirQuality = 1,
   AppUsage = 2, # Once every 30 minutes
   Bluetooth = 60,  # Once per minute. Can have multiple measurements.
   Gyroscope = 720, # Once per 5 seconds. Can have multiple measurements.
   Light = 360, # Once per 10 seconds
   Location = 60, # Once per 60 seconds
   Memory = 60, # Once per minute
   Noise = 120,
   Pedometer = 1,
   Weather = 1,
   Wifi = 60 # once per minute
)
coverage(db = db,
         participant_id = "12345",
         sensor = c("Accelerometer", "Gyroscope"),
         frequency = CARP::freq,
         startDate = "2021-01-01",
         endDate = "2021-05-01")

#> Error in coverage(db = db, participant_id = "12345", sensor = c("Accelerometer",     "Gyroscope"), frequency = CARP::freq, startDate = "2021-01-01",     endDate = "2021-05-01"): object 'db' not found